+++ consumer/config.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/rabbitmq/amqp091-go"
 
-	"github.com/nano-interactive/go-amqp"
-	"github.com/nano-interactive/go-amqp/connection"
-	"github.com/nano-interactive/go-amqp/serializer"
+	"github.com/nano-interactive/go-amqp/v2/connection"
+	"github.com/nano-interactive/go-amqp/v2/logging"
+	"github.com/nano-interactive/go-amqp/v2/serializer"
 )
 
 type ExchangeBinding struct {
@@ -26,7 +26,7 @@ type QueueDeclare struct {
 
 type Config[T any] struct {
 	ctx               context.Context
-	logger            amqp.Logger
+	logger            logging.Logger
 	serializer        serializer.Serializer[T]
 	onError           connection.OnErrorFunc
 	onMessageError    func(context.Context, *amqp091.Delivery, error)
@@ -71,11 +71,15 @@ func WithOnMessageError[T any](onMessageError func(context.Context, *amqp091.Del
 
 func WithQueueConfig[T any](cfg QueueConfig) Option[T] {
 	return func(c *Config[T]) {
+		if cfg.Workers == 0 {
+			cfg.Workers = 1
+		}
+
 		c.queueConfig = cfg
 	}
 }
 
-func WithLogger[T any](logger amqp.Logger) Option[T] {
+func WithLogger[T any](logger logging.Logger) Option[T] {
 	return func(c *Config[T]) {
 		c.logger = logger
 	}
@@ -95,6 +99,38 @@ func WithContext[T any](ctx context.Context) Option[T] {
 
 func WithConnectionOptions[T any](connectionOptions connection.Config) Option[T] {
 	return func(c *Config[T]) {
+		if connectionOptions.Channels == 0 {
+			connectionOptions.Channels = connection.DefaultConfig.Channels
+		}
+
+		if connectionOptions.Vhost == "" {
+			connectionOptions.Vhost = connection.DefaultConfig.Vhost
+		}
+
+		if connectionOptions.Host == "" {
+			connectionOptions.Host = connection.DefaultConfig.Host
+		}
+
+		if connectionOptions.Port == 0 {
+			connectionOptions.Port = connection.DefaultConfig.Port
+		}
+
+		if connectionOptions.User == "" {
+			connectionOptions.User = connection.DefaultConfig.User
+		}
+
+		if connectionOptions.Password == "" {
+			connectionOptions.Password = connection.DefaultConfig.Password
+		}
+
+		if connectionOptions.ReconnectInterval== 0 {
+			connectionOptions.ReconnectInterval = connection.DefaultConfig.ReconnectInterval
+		}
+
+		if connectionOptions.ReconnectRetry == 0 {
+			connectionOptions.ReconnectRetry = connection.DefaultConfig.ReconnectRetry
+		}
+
 		c.connectionOptions = connectionOptions
 	}
 }